refactor(chat-support): use vendor-neutral naming and inline into user service
Rename the Pylon-specific config block, constants and claim struct to
vendor-neutral chat-support naming, and move the token-minting method into
user.go alongside the rest of UserService.

Config moves from pylon.* to chat_support.*, so the env vars become
FLEXPRICE_CHAT_SUPPORT_APP_ID and FLEXPRICE_CHAT_SUPPORT_IDENTITY_SECRET.
Neither key has been set in any environment yet, so no deploy coordination
is needed.

The claim JSON keys (email, name, account_external_id, contact_external_id
and the registered aud/iat/exp) stay as-is: they are the vendor's wire
contract, not internal naming.
