+++ internal/config/config.go
@@ -74,10 +74,10 @@ type Configuration struct {
 	Gemini                     GeminiConfig                     `mapstructure:"gemini" validate:"omitempty"`
 	Whop                       WhopConfig                       `mapstructure:"whop" validate:"omitempty"`
 	Onboarding                 OnboardingConfig                 `mapstructure:"onboarding" validate:"omitempty"`
-	Pylon                      PylonConfig                      `mapstructure:"pylon" validate:"omitempty"`
+	ChatSupport                ChatSupportConfig                `mapstructure:"chat_support" validate:"omitempty"`
 }
 
-type PylonConfig struct {
+type ChatSupportConfig struct {
 	AppID          string `mapstructure:"app_id"`
 	IdentitySecret string `mapstructure:"identity_secret"`
 }
